+++ server/src/index.ts
@@ -8,24 +8,26 @@ import { createUserStrategyConfig } from './config/createUserStrategyConfig';
 import app from './app';
 import { prisma } from './utils/prisma';
 
-const { PORT } = config;
+app();
 
-(async () => {
-  passport.use(await createUserStrategyConfig());
-  passport.serializeUser((user, done) => done(null, user));
-  passport.deserializeUser(async (email: string, done) => {
-    const user = (await prisma.user.findMany()).filter(
-      (user) => user.email === email
-    )[0];
+// const { PORT } = config;
 
-    if (user) return done(null, user);
-    else return done(null, { message: 'User not found' });
-  });
+// (async () => {
+//   passport.use(await createUserStrategyConfig());
+//   passport.serializeUser((user, done) => done(null, user));
+//   passport.deserializeUser(async (email: string, done) => {
+//     const user = (await prisma.user.findMany()).filter(
+//       (user) => user.email === email
+//     )[0];
 
-  app.listen(PORT, () => {
-    console.log(`Server running on port ${PORT}`);
-  });
-})();
+//     if (user) return done(null, user);
+//     else return done(null, { message: 'User not found' });
+//   });
+
+//   app.listen(PORT, () => {
+//     console.log(`Server running on port ${PORT}`);
+//   });
+// })();
 
 // (async () => {
 //   const url = config.DATABASE_URL;
