+++ server/src/routes/loginRoute.ts
@@ -1,5 +1,6 @@
 import express from 'express';
 import { login } from '../controllers/auth/login';
+import passport from 'passport';
 
 const router = express.Router();
